+++ src/front/js/layout.js
@@ -5,11 +5,11 @@ import { BackendURL } from "./component/backendURL";
 
 import { Home } from "./pages/home";
 // import { Member } from "./pages/member";
-import { Single } from "./pages/single";
 import { Footer } from "./component/footer";
 import injectContext from "./store/appContext";
 import { MembershipViewPage } from "./pages/membershipView";
 import { Navbar } from "./component/navbar";
+import { AboutUs } from "./pages/aboutUs";
 
 //create your first component
 const Layout = () => {
@@ -26,6 +26,7 @@ const Layout = () => {
                     <Navbar />
                     <Routes>
                         <Route element={<Home />} path="/" />
+                        <Route element={<AboutUs />} path="/about" />
                         {/* <Route element={<Member />} path="/member" /> */}
                         <Route element={<Single />} path="/single/:theid" />
                         <Route element={< MembershipViewPage/> } path="/membershipView/:id"/>
