+++ src/front/js/pages/aboutUs.js
@@ -1,0 +1,43 @@
+import React from "react";
+import "../../styles/index.css"
+
+export const AboutUs = ()=>{
+    return (
+        <div>
+
+        <div>
+        <h1>Meet the Team</h1>
+
+        {/* <div className="max-w-sm rounded overflow-hidden shadow-lg">
+  <img className="w-full" src="https://via.placeholder.com/350x150" alt="Card image cap" />
+  <div className="px-6 py-4">
+    <div className="font-bold text-xl mb-2">Card Title</div>
+    <p className="text-gray-700 text-base">
+      Lorem ipsum dolor sit amet, consectetur adipisicing elit. Voluptatibus quia, nulla! Maiores et perferendis
+      eaque, exercitationem praesentium nihil.
+    </p>
+  </div>
+</div> */}
+
+        
+        
+        </div>
+
+
+        <h2>Our Mission</h2>
+        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Blanditiis, ipsa? At sequi voluptates fuga dolorem rerum iste! Ea deleniti nam sapiente placeat magnam eaque, harum hic expedita cum culpa. Quis.</p>
+
+        <h2>Locations</h2>
+
+        <address>123 cardio ave. chicago,IL,60456</address>
+        <p>Phone:</p>
+        
+        
+        </div>
+
+
+
+    );
+};
+
+export default AboutUs;
