--- src/front/js/pages/aboutUs.jsx
@@ -1,40 +1,0 @@
-import React, { useContext } from "react";
-import { Context } from "../store/appContext";
-
-export const AboutUs = ()=>{
-    return (
-        <div>
-
-        <h1>Meet the Team</h1>
-        <img src="" alt="photo" />
-        <p>Role:</p>
-        <p>Bio: Lorem ipsum dolor sit amet consectetur adipisicing elit. Enim ipsa quisquam, ex voluptatibus dolore ut cumque. </p>
-        <p>Socials:</p>
-        <img src="" alt="photo" />
-        <p>Role:</p>
-        <p>Bio: Lorem ipsum dolor sit amet consectetur adipisicing elit. Enim ipsa quisquam, ex voluptatibus dolore ut cumque. </p>
-        <p>Socials:</p>
-        <img src="" alt="photo" />
-        <p>Role:</p>
-        <p>Bio: Lorem ipsum dolor sit amet consectetur adipisicing elit. Enim ipsa quisquam, ex voluptatibus dolore ut cumque. </p>
-        <p>Socials:</p>
-        <img src="" alt="photo" />
-        <p>Role:</p>
-        <p>Bio: Lorem ipsum dolor sit amet consectetur adipisicing elit. Enim ipsa quisquam, ex voluptatibus dolore ut cumque. </p>
-        <p>Socials:</p>
-
-        <h2>Our Mission</h2>
-        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Blanditiis, ipsa? At sequi voluptates fuga dolorem rerum iste! Ea deleniti nam sapiente placeat magnam eaque, harum hic expedita cum culpa. Quis.</p>
-
-        <h2>Locations</h2>
-
-        <address>123 cardio ave. chicago,IL,60456</address>
-        <p>Phone:</p>
-        
-        
-        </div>
-
-
-
-    );
-};
